+++ Clauses/Clause.cs
@@ -1,3 +1,4 @@
+using Lex.Expressions;
 using Lex.Tokens;
 
 namespace Lex.Clauses;
@@ -16,6 +17,13 @@ public class Clause
 
     /// <summary>
     /// This property provides the list of tokens that matched the clause that was parsed.
+    /// If this is present, then <see cref="Expressions"/> will not be.
     /// </summary>
     public List<Token> Tokens { get; init; }
+
+    /// <summary>
+    /// This property provide the list of expressions the clause parser captured.  If this
+    /// is present, then <see cref="Tokens"/> will not be.
+    /// </summary>
+    public List<IExpressionTerm> Expressions { get; init; }
 }