+++ docs/release-notes.md
@@ -1,5 +1,12 @@
 ## Release Notes
 
+### 1.1.0
+
+- Added support for an expression clause, allowing expressions to be expected in the
+  midst of other clauses.
+
+- Addition of the [Use Cases](use-cases.md) page to the documentation.
+
 ### 1.0.0
 
 The initial version of the library, at least from the point of view of C#; see